+++ pytim/__init__.py
@@ -1,7 +1,7 @@
 # -*- Mode: python; tab-width: 4; indent-tabs-mode:nil; coding: utf-8 -*-
 # vim: tabstop=4 expandtab shiftwidth=4 softtabstop=4
 #from pytim.patches import PatchTrajectory, PatchOpenMM, PatchMDTRAJ
-from .Interface import SimpleInterface
+from .interface import SimpleInterface
 from .itim import ITIM
 from .gitim import GITIM
 from .willard_chandler import WillardChandler
